:godmode: Beginnings of GameHandler.js
Adds functionality to join a gameroom.
This is probably deprecated before implementation.
Plans to move towards class-based responsibility.
This requires io and socket to be passed towards class.
But since game is a shared object, this might actually make more sense (instead of making more class instances for one game)